+++ RELEASE_NOTES.md
@@ -1,8 +1,13 @@
 # Release Notes
 
-## Next Release - //
+## XMLUnit.NET Java 2.1.1 - //
 
-## XMLUnit for Java 2.1.0 - /Released 2016-03-26/
+* `CompareConstraint` and `ValidationConstraint` for NUnit2 threw
+  `NullReferenceException`s when combined with another failing
+  `Constraint`.
+  Issue similar to [xmlunit/#81](https://github.com/xmlunit/xmlunit/issues/81).
+
+## XMLUnit.NET Java 2.1.0 - /Released 2016-03-26/
 
 * added `CompareConstraint.WithNamespaceContext`
   port of PR [#54](https://github.com/xmlunit/xmlunit/pull/54) by